+++ libs/balances-and-allowances/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [3.15.0](https://github.com/cowprotocol/cowswap/compare/balances-and-allowances-v3.14.0...balances-and-allowances-v3.15.0) (2026-09-23)
+
+### Minor Changes
+
+- feat: show twaps in explorer (#8129)
+
+- feat(solana): apply user slippage from settings (#8164)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/common-utils@3.17.0
+  - @cowprotocol/common-const@3.13.0
+  - @cowprotocol/common-hooks@3.13.0
+  - @cowprotocol/core@3.10.0
+  - @cowprotocol/currency@1.6.0
+  - @cowprotocol/tokens@3.13.0
+  - @cowprotocol/types@4.11.0
+  - @cowprotocol/wallet@3.14.0
+  - @cowprotocol/cowswap-abis@4.3.0
+
 ## [3.14.0](https://github.com/cowprotocol/cowswap/compare/balances-and-allowances-v3.13.1...balances-and-allowances-v3.14.0) (2026-09-15)
 
 ### Minor Changes