+++ libs/core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [3.10.0](https://github.com/cowprotocol/cowswap/compare/core-v3.9.2...core-v3.10.0) (2026-09-23)
+
+### Minor Changes
+
+- feat: show twaps in explorer (#8129)
+
+- feat(solana): apply user slippage from settings (#8164)
+
+### Patch Changes
+
+- fix(rwa): update ondo tokens (#8182)
+
+- Updated dependencies []:
+  - @cowprotocol/common-utils@3.17.0
+  - @cowprotocol/common-const@3.13.0
+
 ## [3.9.2](https://github.com/cowprotocol/cowswap/compare/core-v3.9.1...core-v3.9.2) (2026-09-15)
 
 ### Patch Changes