+++ libs/events/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [4.13.0](https://github.com/cowprotocol/cowswap/compare/events-v4.12.0...events-v4.13.0) (2026-09-23)
+
+### Minor Changes
+
+- feat: show twaps in explorer (#8129)
+
+- feat(solana): apply user slippage from settings (#8164)
+
+- feat: add new prop for TWAP for EOA analytics (#8201)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/types@4.11.0
+
 ## [4.12.0](https://github.com/cowprotocol/cowswap/compare/events-v4.11.0...events-v4.12.0) (2026-09-09)
 
 ### Minor Changes